Fix import in tests. Move shared pytest fixtures to conftest.py.
