+++ pyxem/tests/test_utils/test_indexation_utils.py
@@ -17,53 +17,10 @@
 # along with pyXem.  If not, see <http://www.gnu.org/licenses/>.
 
 import numpy as np
-import pytest
 
-from transforms3d.euler import euler2mat
-import diffpy.structure
-
-from pyxem.libraries.vector_library import DiffractionVectorLibrary
-from pyxem.signals.diffraction_vectors import DiffractionVectors
-from pyxem.utils.indexation_utils import crystal_from_template_matching, \
-    crystal_from_vector_matching, match_vectors
-
-
-@pytest.fixture
-def sp_template_match_result():
-    row_1 = np.array([0, np.array([2, 3, 4]), 0.7], dtype='object')
-    row_2 = np.array([0, np.array([2, 3, 5]), 0.6], dtype='object')
-    # note we require (correlation of row_1 > correlation row_2)
-    return np.vstack((row_1, row_2))
-
-
-@pytest.fixture
-def dp_template_match_result():
-    row_1 = np.array([0, np.array([2, 3, 4]), 0.7], dtype='object')
-    row_2 = np.array([0, np.array([2, 3, 5]), 0.8], dtype='object')
-    row_3 = np.array([1, np.array([2, 3, 4]), 0.5], dtype='object')
-    row_4 = np.array([1, np.array([2, 3, 5]), 0.3], dtype='object')
-    return np.vstack((row_1, row_2, row_3, row_4))
-
-
-@pytest.fixture
-def sp_vector_match_result():
-    # We require (total_error of row_1 > correlation row_2)
-    return np.array([
-        # [phase, R, match_rate, ehkls, total_error]
-        np.array([0, euler2mat(*np.deg2rad([0, 0, 90]), 'rzxz'), 0.5, np.array([0.1, 0.05, 0.2]), 0.1], dtype='object'),
-        np.array([0, euler2mat(*np.deg2rad([0, 0, 90]), 'rzxz'), 0.6, np.array([0.1, 0.1, 0.2]), 0.2], dtype='object')
-    ], dtype='object')
-
-
-@pytest.fixture
-def dp_vector_match_result():
-    return np.array([
-        # [phase, R, match_rate, ehkls, total_error]
-        np.array([0, euler2mat(*np.deg2rad([90, 0, 0]), 'rzxz'), 0.6, np.array([0.1, 0.1, 0.2]), 0.3], dtype='object'),
-        np.array([0, euler2mat(*np.deg2rad([0, 10, 20]), 'rzxz'), 0.5, np.array([0.1, 0.05, 0.2]), 0.4], dtype='object'),
-        np.array([1, euler2mat(*np.deg2rad([0, 45, 45]), 'rzxz'), 0.8, np.array([0.1, 0.3, 0.2]), 0.1], dtype='object'),
-        np.array([1, euler2mat(*np.deg2rad([0, 0, 90]), 'rzxz'), 0.7, np.array([0.1, 0.05, 0.1]), 0.2], dtype='object')
-    ], dtype='object')
+from pyxem.utils.indexation_utils import (crystal_from_template_matching,
+                                          crystal_from_vector_matching,
+                                          match_vectors)
 
 
 def test_crystal_from_template_matching_sp(sp_template_match_result):
@@ -110,29 +67,6 @@ def test_crystal_from_vector_matching_dp(dp_vector_match_result):
     np.testing.assert_allclose(cmap[2]['total_error'], 0.1)
     np.testing.assert_allclose(cmap[2]['orientation_reliability'], r_or)
     np.testing.assert_allclose(cmap[2]['phase_reliability'], r_ph)
-
-
-@pytest.fixture
-def vector_match_peaks():
-    return np.array([
-        [1, 0.1, 0],
-        [0, 2, 0],
-        [1, 2, 3],
-    ])
-
-
-@pytest.fixture
-def vector_library():
-    library = DiffractionVectorLibrary()
-    library['A'] = np.array([
-        [np.array([1, 0, 0]), np.array([0, 2, 0]), 2, 1, np.pi / 2],
-        [np.array([0, 0, 1]), np.array([2, 0, 0]), 1, 2, np.pi / 2],
-    ])
-    lattice = diffpy.structure.Lattice(1, 1, 1, 90, 90, 90)
-    library.structures = [
-        diffpy.structure.Structure(lattice=lattice)
-    ]
-    return library
 
 
 def test_match_vectors(vector_match_peaks, vector_library):
